Add beginnings of arguments!
This parses definitions of optional arguments ("[<foo>]"),
required arguments ("<foo>"), optional trailing globs ("[<foo>[ <foo>[ ..]]]"),
and required globs ("<foo>[ <foo>[ ..]]").

It displays the help for them, but does not further processing.

This does not enforce syntax during the definition pass with regard
to the order that the elements are defined in, but expects that all
options are defined before all arguments.

(note this means that a glob can occur before a non glob which
doesn't make sense @todo.)

we just decided that this must change.